argument

+++ Assets/JunMo/Argument/SkillArgumentData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "SkillArgumentData", menuName = "Scriptable Objects/SkillArgumentData")]
 public class SkillArgumentData : ArgumentData
 {
-    public float maxLevel = 5;
+    public int maxLevel = 5;
     public int level = 1;
     public SkillType skillType;
     public float[] levelValues;
